+++ apps/web/app/game/crash/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { LiveResults } from "@winrlabs/web3-games";
 import dynamic from "next/dynamic";
 const CDN_URL = process.env.NEXT_PUBLIC_BASE_CDN_URL || "";
 
@@ -12,17 +13,20 @@ const CrashGame = dynamic(
 
 const CrashPage = () => {
   return (
-    <CrashGame
-      minWager={0.1}
-      maxWager={2000}
-      gameUrl={`${CDN_URL}/builded-games/crash`}
-      options={{
-        scene: {
-          loader: "/crash/loader.png",
-          logo: "/crash/game-logo.svg",
-        },
-      }}
-    />
+    <>
+      <CrashGame
+        minWager={0.1}
+        maxWager={2000}
+        gameUrl={`${CDN_URL}/builded-games/crash`}
+        options={{
+          scene: {
+            loader: "/crash/loader.png",
+            logo: "/crash/game-logo.svg",
+          },
+        }}
+      />
+      <LiveResults />
+    </>
   );
 };