+++ apps/web/app/game/plinko-2d/page.tsx
@@ -1,17 +1,20 @@
 "use client";
 
-import { PlinkoGame } from "@winrlabs/web3-games";
+import { LiveResults, PlinkoGame } from "@winrlabs/web3-games";
 
 export default function PlinkoPage() {
   return (
-    <PlinkoGame
-      options={{
-        scene: {
-          backgroundImage: "url(/coin-flip/coin-flip-bg.png)",
-        },
-      }}
-      minWager={0.1}
-      maxWager={2000}
-    />
+    <>
+      <PlinkoGame
+        options={{
+          scene: {
+            backgroundImage: "url(/coin-flip/coin-flip-bg.png)",
+          },
+        }}
+        minWager={0.1}
+        maxWager={2000}
+      />
+      <LiveResults />
+    </>
   );
 }