+++ apps/web/app/game/wheel/page.tsx
@@ -1,7 +1,12 @@
 "use client";
 
-import { WheelGame } from "@winrlabs/web3-games";
+import { LiveResults, WheelGame } from "@winrlabs/web3-games";
 
 export default function WheelPage() {
-  return <WheelGame minWager={2} maxWager={2000} options={{}} />;
+  return (
+    <>
+      <WheelGame minWager={2} maxWager={2000} options={{}} />;
+      <LiveResults />
+    </>
+  );
 }