+++ packages/games/src/lib/solo/mines/components/template.tsx
@@ -1,17 +1,18 @@
 import { zodResolver } from "@hookform/resolvers/zod";
+import debounce from "debounce";
 import React from "react";
-import { useForm, useFormContext } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import z from "zod";
-import { Mines } from "..";
+
 import { GameContainer, SceneContainer } from "../../../common/containers";
 import { Form } from "../../../ui/form";
 import { toDecimals } from "../../../utils/web3";
+import { Mines } from "..";
 import { initialBoard } from "../constants";
 import mineMultipliers from "../constants/mines-multipliers.json";
 import { useMinesGameStateStore } from "../store";
-import { FormSetValue, MinesFormField } from "../types";
+import { FormSetValue, MINES_GAME_STATUS, MinesFormField } from "../types";
 import { MinesGameProps } from "./game";
-import debounce from "debounce";
 
 type TemplateProps = MinesGameProps & {
   minWager?: number;
@@ -22,7 +23,7 @@ type TemplateProps = MinesGameProps & {
 };
 
 const MinesTemplate = ({ ...props }: TemplateProps) => {
-  const { board } = useMinesGameStateStore(["board"]);
+  const { board, gameStatus } = useMinesGameStateStore(["board", "gameStatus"]);
 
   const formSchema = z.object({
     wager: z
@@ -100,6 +101,23 @@ const MinesTemplate = ({ ...props }: TemplateProps) => {
     form.setValue(props.formSetValue.key, props.formSetValue.value);
   }, [props.formSetValue]);
 
+  React.useEffect(() => {
+    const values = form.getValues();
+
+    if (
+      values.selectedCells.some(
+        (c) => c === true && gameStatus !== MINES_GAME_STATUS.ENDED
+      )
+    )
+      props.onSubmitGameForm(values);
+  }, [form.getValues("selectedCells")]);
+
+  React.useEffect(() => {
+    if (gameStatus == MINES_GAME_STATUS.ENDED) {
+      props.onAnimationCompleted && props.onAnimationCompleted([] as any);
+    }
+  }, [gameStatus]);
+
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(props.onSubmitGameForm)}>
@@ -113,7 +131,10 @@ const MinesTemplate = ({ ...props }: TemplateProps) => {
               currentMultiplier={currentMultiplier}
             />
             <SceneContainer className="lg:wr-h-[790px]">
-              <Mines.Scene currentMultiplier={currentMultiplier} />
+              <Mines.Scene
+                currentMultiplier={currentMultiplier}
+                isLoading={props.isLoading}
+              />
             </SceneContainer>
           </Mines.Game>
         </GameContainer>