+++ packages/api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @winrlabs/api
 
+## 1.2.0
+
+### Minor Changes
+
+- cfca46d: mines ux update
+
+## 1.1.1
+
+### Patch Changes
+
+- 94adbd1: bump
+
+## 1.1.0
+
+### Minor Changes
+
+- af62052: Constant url
+
 ## 1.0.1
 
 ### Patch Changes