+++ packages/games/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @winrlabs/games
 
+## 2.11.0
+
+### Minor Changes
+
+- cfca46d: mines ux update
+- 442a4f8: live results
+
+## 2.10.1
+
+### Patch Changes
+
+- 94adbd1: bump
+
+## 2.10.0
+
+### Minor Changes
+
+- af62052: Constant url
+
 ## 2.9.0
 
 ### Minor Changes