+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @winrlabs/ui
 
+## 3.0.0
+
+### Minor Changes
+
+- cfca46d: mines ux update
+
+### Patch Changes
+
+- Updated dependencies [cfca46d]
+- Updated dependencies [442a4f8]
+  - @winrlabs/web3@0.4.0
+
+## 2.0.1
+
+### Patch Changes
+
+- 94adbd1: bump
+- Updated dependencies [94adbd1]
+  - @winrlabs/web3@0.3.1
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [af62052]
+  - @winrlabs/web3@0.3.0
+
 ## 1.0.0
 
 ### Patch Changes