+++ packages/web3-games/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @winrlabs/web3-games
 
+## 2.0.0
+
+### Minor Changes
+
+- cfca46d: mines ux update
+- 442a4f8: live results
+
+### Patch Changes
+
+- Updated dependencies [cfca46d]
+- Updated dependencies [442a4f8]
+  - @winrlabs/api@1.2.0
+  - @winrlabs/games@2.11.0
+  - @winrlabs/web3@0.4.0
+  - @winrlabs/web3-games@2.0.0
+
+## 1.0.1
+
+### Patch Changes
+
+- 94adbd1: bump
+- Updated dependencies [94adbd1]
+  - @winrlabs/api@1.1.1
+  - @winrlabs/games@2.10.1
+  - @winrlabs/web3@0.3.1
+  - @winrlabs/web3-games@1.0.1
+
+## 1.0.0
+
+### Minor Changes
+
+- af62052: Constant url
+
+### Patch Changes
+
+- Updated dependencies [af62052]
+  - @winrlabs/api@1.1.0
+  - @winrlabs/games@2.10.0
+  - @winrlabs/web3@0.3.0
+  - @winrlabs/web3-games@1.0.0
+
 ## 1.0.0
 
 ### Minor Changes