+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @winrlabs/web3
 
+## 0.4.0
+
+### Minor Changes
+
+- cfca46d: mines ux update
+- 442a4f8: live results
+
+### Patch Changes
+
+- Updated dependencies [cfca46d]
+  - @winrlabs/api@1.2.0
+
+## 0.3.1
+
+### Patch Changes
+
+- 94adbd1: bump
+- Updated dependencies [94adbd1]
+  - @winrlabs/api@1.1.1
+
+## 0.3.0
+
+### Minor Changes
+
+- af62052: Constant url
+
+### Patch Changes
+
+- Updated dependencies [af62052]
+  - @winrlabs/api@1.1.0
+
 ## 0.1.0
 
 ### Minor Changes